moves parameter bag code to its own file
