+++ parameter_bag.go
@@ -1,0 +1,45 @@
+package http_router
+
+const (
+	ParamsBagKey = "urlParameters"
+)
+
+type urlParameter struct {
+	name  string
+	value string
+}
+
+type UrlParameterBag struct {
+	params []urlParameter
+}
+
+func (u *UrlParameterBag) addParameter(param urlParameter) {
+	if u.params == nil {
+		u.params = make([]urlParameter, 0, 5)
+	}
+
+	u.params = append(u.params, param)
+}
+
+func (u *UrlParameterBag) GetByName(name string, def string) string {
+	for _, item := range u.params {
+		if item.name == name {
+			return item.value
+		}
+	}
+
+	return def
+}
+
+func (u *UrlParameterBag) GetByIndex(index uint, def string) string {
+	i := int(index)
+	if len(u.params) <= i {
+		return def
+	}
+
+	return u.params[i].value
+}
+
+func NewUrlParameterBag() UrlParameterBag {
+	return UrlParameterBag{}
+}
