+++ router.go
@@ -68,10 +68,6 @@ type PrefixTreeRouter struct {
 	tree Tree
 }
 
-const (
-	ParamsBagKey = "urlParameters"
-)
-
 func (r *PrefixTreeRouter) ServeHTTP(response http.ResponseWriter, request *http.Request) {
 	handler, params := r.tree.Find(request.Method, request.URL.Path)
 	if handler == nil {
@@ -145,44 +141,4 @@ func (r *PrefixTreeRouter) AddHandler(verb, path string, handler HandlerFunction
 	}
 
 	r.tree.Insert(verb, parser.chunks, handler)
-}
-
-type urlParameter struct {
-	name  string
-	value string
-}
-
-type UrlParameterBag struct {
-	params []urlParameter
-}
-
-func (u *UrlParameterBag) addParameter(param urlParameter) {
-	if u.params == nil {
-		u.params = make([]urlParameter, 0, 5)
-	}
-
-	u.params = append(u.params, param)
-}
-
-func (u *UrlParameterBag) GetByName(name string, def string) string {
-	for _, item := range u.params {
-		if item.name == name {
-			return item.value
-		}
-	}
-
-	return def
-}
-
-func (u *UrlParameterBag) GetByIndex(index uint, def string) string {
-	i := int(index)
-	if len(u.params) <= i {
-		return def
-	}
-
-	return u.params[i].value
-}
-
-func NewUrlParameterBag() UrlParameterBag {
-	return UrlParameterBag{}
 }
